Made samson base compile in 32 bit. Improvements to makefile. Prepared export_to_mongo.
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@1508 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ apps/samsonPush/main_samsonPush.cpp
@@ -58,7 +58,7 @@ int main( int argc , char *argv[] )
         if( (size + line_length ) > BUFFER_SIZE )
         {
             // Process buffer
-            fprintf(stderr , "Flushing %lu bytes to queue %s\n", size , argv[2] );
+		    fprintf(stderr , "Flushing %lu bytes to queue %s\n", (unsigned long int) size , argv[2]);
             client.push(  argv[2]  , 0 , buffer, size );
             
             // Process buffer
@@ -71,7 +71,7 @@ int main( int argc , char *argv[] )
     }
     
     // Process last buffer
-    fprintf(stderr , "Flushing %lu bytes to queue %s\n", size , argv[2] );
+    fprintf(stderr , "Flushing %lu bytes to queue %s\n", (unsigned long int) size , argv[2] );
     client.push(  argv[2]  , 0 , buffer, size );
     
     // Wait until all operations are complete
